+++ packages/react-sdk/src/core/components/CallLayout/SpeakerLayout.tsx
@@ -72,6 +72,7 @@ export const SpeakerLayout = ({
     if (!participantsBarWrapperElement || !call) return;
 
     const cleanup = call.setViewport(participantsBarWrapperElement);
+
     return () => cleanup();
   }, [participantsBarWrapperElement, call]);