+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,34 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.1.4](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.1.3...@stream-io/video-react-native-sdk-0.1.4) (2023-10-09)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.3.31`
+* `@stream-io/video-react-bindings` updated to version `0.2.32`
+
+### Features
+
+* **react-native:** add all push support to Expo and non ringing push for vanilla  ([#1097](https://github.com/GetStream/stream-video-js/issues/1097)) ([9dcbe23](https://github.com/GetStream/stream-video-js/commit/9dcbe23dc949e452132b5450419a9558dc836309))
+
+### [0.1.3](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.1.2...@stream-io/video-react-native-sdk-0.1.3) (2023-10-06)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.3.30`
+* `@stream-io/video-react-bindings` updated to version `0.2.31`
+
+### Features
+
+* ScreenShare Audio support ([#1118](https://github.com/GetStream/stream-video-js/issues/1118)) ([5b63e1c](https://github.com/GetStream/stream-video-js/commit/5b63e1c5f52c76e3761e6907bd3786c19f0e5c6d))
+
+### [0.1.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.1.1...@stream-io/video-react-native-sdk-0.1.2) (2023-10-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.3.29`
+* `@stream-io/video-react-bindings` updated to version `0.2.30`
 ### [0.1.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.1.0...@stream-io/video-react-native-sdk-0.1.1) (2023-10-04)